requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch==0.5.0
timm==1.0.26
albumentations==1.4.24
opencv-python-headless
numpy
pandas
pillow
pyyaml
scikit-learn
scipy
tqdm
matplotlib
datasets>=2.14.0
seaborn
nibabel
rich
pydicom
kaggle==1.7.4.5
